requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
openai>=1.50
aiofiles>=23.2
python-multipart>=0.0.9
